home-encrypted-backup-doc-system

+++ Edit_Start.command
@@ -11,12 +11,15 @@ then
     mkdir encrypt_dir
 else
     openssl aes-256-cbc -d -in encrypt_dir.aes -out encrypt_dir.tar.gz
-    rm -r encrypt_dir.aes
     tar -zxvf encrypt_dir.tar.gz
     rm -r encrypt_dir.tar.gz
 fi
 
-mv .Edit_Stop.command Edit_Stop.command
-mv Edit_Start.command .Edit_Start.command
-mv Read_Start.command .Read_Start.command
-mv Pull_Sync.command .Pull_Sync.command
+if [ -d encrypt_dir ]
+then
+    rm -r encrypt_dir.aes
+    mv .Edit_Stop.command Edit_Stop.command
+    mv Edit_Start.command .Edit_Start.command
+    mv Read_Start.command .Read_Start.command
+    mv Pull_Sync.command .Pull_Sync.command
+fi
